+++ poly/src/lib.rs
@@ -23,10 +23,12 @@ pub mod polynomial;
 pub use domain::{
     EvaluationDomain, GeneralEvaluationDomain, MixedRadixEvaluationDomain, Radix2EvaluationDomain,
 };
-pub use evaluations::multivariate::multilinear::{
-    DenseMultilinearExtension, MultilinearExtension, SparseMultilinearExtension,
+pub use evaluations::{
+    multivariate::multilinear::{
+        DenseMultilinearExtension, MultilinearExtension, SparseMultilinearExtension,
+    },
+    univariate::Evaluations,
 };
-pub use evaluations::univariate::Evaluations;
 pub use polynomial::{multivariate, univariate, MVPolynomial, Polynomial, UVPolynomial};
 
 #[cfg(test)]